backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
apscheduler==3.10.4
psutil==5.9.8
chromadb==0.4.24
duckdb==0.10.1
cryptography==42.0.7
pyyaml==6.0.1
python-dotenv==1.0.1
